refactor: move inline tests to plugin_tests.rs and enhance detect_is_anthropic
- Move #[cfg(test)] mod tests from deepseek/plugin.rs to separate
  deepseek/plugin_tests.rs following CONTRIBUTING.md convention
- Add #[path = "plugin_tests.rs"] mod plugin_tests; in plugin.rs
- Enhance detect_is_anthropic() with secondary heuristic: top-level
  'system' field is exclusive to Anthropic protocol format
- All 457 closeclaw-llm tests pass

Source: user
Type: refactor
